remove separate format file
